requirements.txt
Appium-Python-Client>=2.0.0,<3.0.0